angepasste Logik

+++ TEST/Assets/GameAssets/Skripte/Healthbar/HealthbarSkript.cs
@@ -7,12 +7,12 @@ public class HealthbarSkript : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
     }
 
-    public void SetMaxHealth(int maxhealth)
+    public void SetMaxHealth(float maxhealth)
     {
         slider.maxValue = maxhealth;
         slider.value = maxhealth;
